+++ e2e/node/canisters/counter.ts
@@ -62,7 +62,6 @@ export const createActor = async (options?: HttpAgentOptions, agent?: Agent) => 
   }
 
   const canisterId = await Actor.createCanister({ agent: effectiveAgent });
-  console.log('Installing canister with ID:', canisterId.toText());
   await Actor.install({ module }, { canisterId, agent: effectiveAgent });
   return Actor.createActor(idl, { canisterId, agent: effectiveAgent }) as ActorSubclass<_SERVICE>;
 };